+++ src/app/netsuite/netsuite.component.html
@@ -1,6 +1,6 @@
 <mat-sidenav-container class="netsuite" autosize fxFlexFill *ngIf="!isLoading">
     <mat-sidenav class="netsuite-sidenav" mode="side" opened>
-      <mat-nav-list>
+      <mat-nav-list class="mat-drawer-inner">
         <header matSubheader class="netsuite-sidenav--header">
           <img alt="Fyle" src="../../assets/images/pngs/fyle-logo.png" class="netsuite-sidenav--logo" />
           <mat-icon>add</mat-icon>
@@ -25,12 +25,12 @@
         <a class="mat-h3 netsuite-sidenav--nav-item" mat-list-item>
           Settings
         </a>
-        <a class="mat-h3 netsuite-sidenav--nav-sub-item" mat-list-item *ngIf="connectNetSuite"
-        routerLink="{{workspace.id}}/settings/netsuite/connect/" routerLinkActive="active">
-        Connect to NetSuite
+        <a class="mat-h3 netsuite-sidenav--nav-sub-item" mat-list-item *ngIf="!netsuiteConnected"
+        routerLink="{{workspace.id}}/settings/netsuite/" routerLinkActive="active">
+          Connect to NetSuite
         </a>
-        <a class="mat-h3 netsuite-sidenav--nav-sub-item" mat-list-item routerLink="{{workspace.id}}/settings/configurations/"
-          routerLinkActive="active">
+        <a class="mat-h3 netsuite-sidenav--nav-sub-item" [ngClass]="{'netsuite-sidenav--nav-sub-item__disabled': navDisabled}" mat-list-item
+         routerLink="{{workspace.id}}/settings/configurations/general/" routerLinkActive="active">
           Configurations
         </a>
         <a class="mat-h3 netsuite-sidenav--nav-sub-item" [ngClass]="{'netsuite-sidenav--nav-sub-item__disabled': navDisabled}" mat-list-item routerLink="{{workspace.id}}/settings/general_mappings/"
