+++ src/app/netsuite/settings/connect-netsuite/connect-netsuite.component.html
@@ -1,20 +1,74 @@
-<div class="internal-loader" fxLayoutAlign="center start" *ngIf="isParentLoading" fxFlexFill>
+<div class="internal-loader" fxLayoutAlign="center start" *ngIf="isLoading" fxFlexFill>
   <app-loader></app-loader>
 </div>
-<div class="page-header" *ngIf="!isParentLoading">
-  <div class="page-header--name mat-h2">
-    NetSuite Connection
-  </div>
-  <nav mat-tab-nav-bar>
-    <a mat-tab-link (click)="changeState('CONNECT')" [active]="state==='CONNECT'">
-      NetSuite
-    </a>
-    <a mat-tab-link (click)="changeState('SUBSIDIARY')" [active]="state==='SUBSIDIARY'" *ngIf='netsuiteConnectionDone'>
-      Subsidiary
-    </a>
-  </nav>
-</div>
 
-<div *ngIf="!isParentLoading">
-  <router-outlet></router-outlet>
+<div class="configurations" *ngIf="!isLoading && !netsuiteConnectionDone">
+  <div class="page-header configurations--header">
+    <div class="page-header--name mat-h2">
+      NetSuite Connection
+    </div>
+  </div>
+  <div class="page-info">
+    <div fxLayout="row" fxLayoutAlign="start center" class="page-info--description">
+      <div>
+        Connect to your NetSuite account to start exporting expense data from Fyle.
+      </div>
+      <a href="" class="read-more-link" fxLayout="row" fxLayoutAlign="start center">Read more<mat-icon class="read-more-link--icon">open_in_new</mat-icon></a>
+    </div>
+  </div>
+  <div *ngIf="!netsuiteConnectionDone">
+  <form class="configurations--form" [formGroup]="connectNetSuiteForm">
+    <div fxLayout="column" fxLayoutAlign="start stretch">
+
+      <div class="mat-body-strong">
+        <strong class="mappings-dialog--label">
+          NetSuite Account ID
+        </strong>
+      </div>
+      <mat-form-field class="mappings-dialog--form-element" appearance="outline">
+        <input required type="text" formControlName="nsAccountId" placeholder="Enter Account ID" matInput>
+      </mat-form-field>
+
+      <div class="mat-body-strong">
+        <strong class="mappings-dialog--label">
+          NetSuite Consumer Key
+        </strong>
+      </div>
+      <mat-form-field class="mappings-dialog--form-element" appearance="outline">
+        <input required type="text" formControlName="nsConsumerKey" placeholder="Enter Consumer Key" matInput>
+      </mat-form-field>
+
+      <div class="mat-body-strong">
+        <strong class="mappings-dialog--label">
+          NetSuite Consumer Secret
+        </strong>
+      </div>
+      <mat-form-field class="mappings-dialog--form-element" appearance="outline">
+        <input required type="text" formControlName="nsConsumerSecret" placeholder="Enter Consumer Secret" matInput>
+      </mat-form-field>
+
+      <div class="mat-body-strong">
+        <strong class="mappings-dialog--label">
+          NetSuite Token ID
+        </strong>
+      </div>
+      <mat-form-field class="mappings-dialog--form-element" appearance="outline">
+        <input required type="text" formControlName="nsTokenId" placeholder="Enter Token ID" matInput>
+      </mat-form-field>
+
+      <div class="mat-body-strong">
+          <strong class="mappings-dialog--label">
+            NetSuite Token Secret
+          </strong>
+        </div>
+        <mat-form-field class="mappings-dialog--form-element" appearance="outline">
+          <input required type="text" formControlName="nsTokenSecret" placeholder="Enter Token Secret" matInput>
+        </mat-form-field>
+
+      <button [disabled]="isSaveDisabled || !connectNetSuiteForm.valid" class="configurations--save-btn" mat-flat-button color="primary" (click)="save()" type="submit">
+        Save
+      </button>
+    </div>
+  </form>
+  </div>
 </div>
