+++ src/app/netsuite/settings/netsuite-configurations/configuration/configuration.component.html
@@ -2,11 +2,6 @@
   <app-loader></app-loader>
 </div>
 <div class="configurations" *ngIf="!isLoading">
-  <div class="page-header configurations--header">
-    <div class="page-header--name mat-h2">
-      Configurations
-    </div>
-  </div>
   <div class="page-info">
     <div fxLayout="row" fxLayoutAlign="start center" class="page-info--description">
       <div>
