+++ src/app/netsuite/settings/netsuite-configurations/configuration/configuration.component.spec.ts
