+++ src/app/netsuite/settings/netsuite-configurations/netsuite-configurations.component.html
@@ -1,0 +1,20 @@
+<div class="internal-loader" fxLayoutAlign="center start" *ngIf="isParentLoading" fxFlexFill>
+    <app-loader></app-loader>
+  </div>
+  <div class="page-header" *ngIf="!isParentLoading">
+    <div class="page-header--name mat-h2">
+      NetSuite Configurations
+    </div>
+    <nav mat-tab-nav-bar>
+      <a mat-tab-link (click)="changeState('SUBSIDIARY')" [active]="state==='SUBSIDIARY'" *ngIf='netsuiteConnectionDone'>
+        Subsidiary
+      </a>
+      <a mat-tab-link (click)="changeState('GENERAL')" [active]="state==='GENERAL'">
+        Configurations
+      </a>
+    </nav>
+  </div>
+  
+  <div *ngIf="!isParentLoading">
+    <router-outlet></router-outlet>
+  </div>
