+++ src/app/netsuite/settings/netsuite-configurations/netsuite-configurations.component.scss
