+++ src/app/netsuite/settings/netsuite-configurations/netsuite-configurations.component.ts
@@ -1,0 +1,54 @@
+import { Component, OnInit } from '@angular/core';
+import { SettingsService } from 'src/app/core/services/settings.service';
+import { MappingsService } from 'src/app/core/services/mappings.service';
+import { ActivatedRoute, Router } from '@angular/router';
+
+@Component({
+  selector: 'app-netsuite-configurations',
+  templateUrl: './netsuite-configurations.component.html',
+  styleUrls: ['./netsuite-configurations.component.scss', '../../netsuite.component.scss']
+})
+export class NetsuiteConfigurationsComponent implements OnInit {
+
+  state: string;
+  workspaceId: number;
+  isParentLoading: boolean;
+  fyleFields: any;
+  netsuiteConnection: any;
+  netsuiteConnectionDone: boolean;
+
+  constructor(private route: ActivatedRoute, private router: Router, private mappingsService: MappingsService, private settingsService: SettingsService) { }
+
+  changeState(state) {
+    const that = this;
+    if (that.state === 'GENERAL') {
+      that.state = state;
+      that.router.navigate([`workspaces/${that.workspaceId}/settings/configurations/${that.state.toLowerCase()}`]);
+    }
+    if (that.state === 'SUBSIDIARY') {
+      that.state = state;
+      that.router.navigate([`workspaces/${that.workspaceId}/settings/configurations/${that.state.toLowerCase()}`]);
+    }
+  }
+
+  ngOnInit() {
+    const that = this;
+    
+    that.isParentLoading = true;
+
+    that.netsuiteConnection = false;
+
+    that.workspaceId = +that.route.parent.snapshot.params.workspace_id;
+
+    that.state = that.route.snapshot.firstChild.routeConfig.path.toUpperCase() || 'SUBSIDIARY';
+    that.settingsService.getNetSuiteCredentials(that.workspaceId).subscribe((response) => {
+      if (response) {
+        that.netsuiteConnectionDone = true;
+      }
+      that.isParentLoading = false;
+    }, () => {
+      that.isParentLoading = false;
+    });
+  }
+
+}
