+++ src/app/netsuite/settings/netsuite-configurations/subsidiary/subsidiary.component.scss
